Move content out to separate content component
